Fixed CAST datatype in ownerfees CLEAN tables (#3928)

+++ bietlejuice/jobs/composer/db/datalake/queries/owner_fees/clean/full/contract_aud.sql
@@ -6,8 +6,8 @@ SELECT
     rev,
     revtype AS rev_type,
     revend AS rev_end,
-    CAST(real_state_agent_share AS DECIMAL(10,2)) AS real_state_agent_share,
-    CAST(rent AS DECIMAL(10,2)) AS rent,
+    CAST(real_state_agent_share AS FLOAT) AS real_state_agent_share,
+    CAST(rent AS FLOAT) AS rent,
     validity_date AS dt_validity,
     created_at AS ts_updated,
     updated_at AS ts_created
