+++ bietlejuice/jobs/composer/db/datalake/queries/owner_fees/clean/full/contract_brokerage_fee_aud.sql
@@ -6,9 +6,9 @@ SELECT
     rev,
     revtype AS rev_type,
     revend AS rev_end,
-    CAST(fee AS DECIMAL(10,2)) AS brokerage_fee,
-    CAST(premium_fee AS DECIMAL(10,2)) AS premium_fee,
-    CAST(down_payment AS DECIMAL(10,2)) AS down_payment,
+    CAST(fee AS FLOAT) AS brokerage_fee,
+    CAST(premium_fee AS FLOAT) AS premium_fee,
+    CAST(down_payment AS FLOAT) AS down_payment,
     installment_number_mod AS mod_installment_number,
     premium_fee_mod AS mod_premium_fee
 FROM
